+++ protobuf/DelGroupActivity/DataReq.java
@@ -6,9 +6,9 @@ import com.squareup.wire.ProtoField;
 public final class DataReq extends Message {
     @ProtoField(tag = 2, type = Message.Datatype.INT32)
     public final Integer activityId;
-    @ProtoField(tag = 1, type = Message.Datatype.INT32)
-    public final Integer groupId;
-    public static final Integer DEFAULT_GROUPID = 0;
+    @ProtoField(tag = 1, type = Message.Datatype.INT64)
+    public final Long groupId;
+    public static final Long DEFAULT_GROUPID = 0L;
     public static final Integer DEFAULT_ACTIVITYID = 0;
 
     private DataReq(Builder builder, boolean z) {
@@ -34,7 +34,7 @@ public final class DataReq extends Message {
     /* loaded from: classes3.dex */
     public static final class Builder extends Message.Builder<DataReq> {
         public Integer activityId;
-        public Integer groupId;
+        public Long groupId;
 
         public Builder() {
         }
