+++ protobuf/DelGroupUsers/DataRes.java
@@ -2,11 +2,11 @@ package protobuf.DelGroupUsers;
 
 import com.squareup.wire.Message;
 import com.squareup.wire.ProtoField;
-/* loaded from: classes.dex */
+/* loaded from: classes2.dex */
 public final class DataRes extends Message {
-    public static final Integer DEFAULT_GROUPID = 0;
-    @ProtoField(tag = 1, type = Message.Datatype.INT32)
-    public final Integer groupId;
+    public static final Long DEFAULT_GROUPID = 0L;
+    @ProtoField(tag = 1, type = Message.Datatype.INT64)
+    public final Long groupId;
 
     private DataRes(Builder builder, boolean z) {
         super(builder);
@@ -22,9 +22,9 @@ public final class DataRes extends Message {
         this.groupId = builder.groupId;
     }
 
-    /* loaded from: classes.dex */
+    /* loaded from: classes2.dex */
     public static final class Builder extends Message.Builder<DataRes> {
-        public Integer groupId;
+        public Long groupId;
 
         public Builder() {
         }
